color picker hex input, app.jsx component, font changes
