+++ color-picker-project/package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "@uiw/react-color": "^2.3.0",
     "@uiw/react-color-alpha": "^2.3.0",
+    "@uiw/react-color-editable-input": "^2.3.0",
     "@uiw/react-color-editable-input-hsla": "^2.3.0",
     "@uiw/react-color-material": "^2.3.0",
     "@uiw/react-color-wheel": "^2.3.0",
